requirements.txt
fastapi
uvicorn
pandas
google-cloud-aiplatform==1.64.0
langchain
langchain-community
langchain-google-community
langchain-google-community[featurestore]
langchain-google-vertexai
google-cloud-storage